# docs/openapi.yaml
openapi: 3.1.0
info:
  title: PinkGlasses API
  version: "1"
  description: |
    Every route the server registers, generated from the router (go run ./tools/openapi).
    Paths, methods, roles and authentication are exact; request and response bodies are
    described in wiki/API.md rather than typed here. Errors are {"error": "<sentence>"}.
servers:
  - url: /
components:
  securitySchemes:
    session:
      type: apiKey
      in: cookie
      name: pg_session
    token:
      type: http
      scheme: bearer
      description: An API token, pgt_… — may be narrower than its owner, never wider.
    proxy:
      type: apiKey
      in: header
      name: X-Forwarded-User
      description: Honoured only alongside X-Proxy-Secret when ASM_TRUSTED_PROXY_SECRET is set.
  responses:
    e400:
      description: malformed or missing input
      content:
        application/json:
          schema:
            type: object
            properties:
              error: {type: string}
    e401:
      description: not signed in, or the credential is no longer valid
      content:
        application/json:
          schema:
            type: object
            properties:
              error: {type: string}
    e403:
      description: signed in, but the route needs a higher role
      content:
        application/json:
          schema:
            type: object
            properties:
              error: {type: string}
    e404:
      description: no such thing
      content:
        application/json:
          schema:
            type: object
            properties:
              error: {type: string}
    e409:
      description: conflicts with state — last administrator, username taken, empty pool
      content:
        application/json:
          schema:
            type: object
            properties:
              error: {type: string}
paths:
  /api/v1/auth/status:
    get:
      operationId: getAuthStatus
      summary: "{setup_required, user?} — whether the install still needs its first administrator, and who you are if signed in"
      x-required-role: public
      security: []
      responses:
        "200": {description: ok}
        "400": {$ref: "#/components/responses/e400"}
        "404": {$ref: "#/components/responses/e404"}
        "409": {$ref: "#/components/responses/e409"}
  /api/v1/auth/setup:
    post:
      operationId: postAuthSetup
      summary: "create the first administrator — {username, display_name, password}; refuses once any account exists"
      x-required-role: public
      security: []
      responses:
        "200": {description: ok}
        "201": {description: created}
        "400": {$ref: "#/components/responses/e400"}
        "404": {$ref: "#/components/responses/e404"}
        "409": {$ref: "#/components/responses/e409"}
  /api/v1/auth/login:
    post:
      operationId: postAuthLogin
      summary: "{username, password} → sets the session cookie, returns {user}"
      x-required-role: public
      security: []
      responses:
        "200": {description: ok}
        "201": {description: created}
        "400": {$ref: "#/components/responses/e400"}
        "404": {$ref: "#/components/responses/e404"}
        "409": {$ref: "#/components/responses/e409"}
  /api/v1/auth/logout:
    post:
      operationId: postAuthLogout
      summary: "end this session"
      x-required-role: any
      security:
        - session: []
        - token: []
        - proxy: []
      responses:
        "200": {description: ok}
        "201": {description: created}
        "400": {$ref: "#/components/responses/e400"}
        "401": {$ref: "#/components/responses/e401"}
        "404": {$ref: "#/components/responses/e404"}
        "409": {$ref: "#/components/responses/e409"}
  /api/v1/auth/me:
    get:
      operationId: getAuthMe
      summary: "{user, via, must_change_password}"
      x-required-role: any
      security:
        - session: []
        - token: []
        - proxy: []
      responses:
        "200": {description: ok}
        "400": {$ref: "#/components/responses/e400"}
        "401": {$ref: "#/components/responses/e401"}
        "404": {$ref: "#/components/responses/e404"}
        "409": {$ref: "#/components/responses/e409"}
  /api/v1/auth/password:
    post:
      operationId: postAuthPassword
      summary: "{current_password, new_password} — signs your other sessions out"
      x-required-role: any
      security:
        - session: []
        - token: []
        - proxy: []
      responses:
        "200": {description: ok}
        "201": {description: created}
        "400": {$ref: "#/components/responses/e400"}
        "401": {$ref: "#/components/responses/e401"}
        "404": {$ref: "#/components/responses/e404"}
        "409": {$ref: "#/components/responses/e409"}
  /api/v1/scopes:
    get:
      operationId: getScopes
      summary: "all companies; ?mine=true narrows to ones you created"
      x-required-role: viewer
      security:
        - session: []
        - token: []
        - proxy: []
      responses:
        "200": {description: ok}
        "400": {$ref: "#/components/responses/e400"}
        "401": {$ref: "#/components/responses/e401"}
        "403": {$ref: "#/components/responses/e403"}
        "404": {$ref: "#/components/responses/e404"}
        "409": {$ref: "#/components/responses/e409"}
    post:
      operationId: postScopes
      summary: "{name}"
      x-required-role: operator
      security:
        - session: []
        - token: []
        - proxy: []
      responses:
        "200": {description: ok}
        "201": {description: created}
        "400": {$ref: "#/components/responses/e400"}
        "401": {$ref: "#/components/responses/e401"}
        "403": {$ref: "#/components/responses/e403"}
        "404": {$ref: "#/components/responses/e404"}
        "409": {$ref: "#/components/responses/e409"}
  /api/v1/scopes/{scopeID}/summary:
    get:
      operationId: getScopesByScopeIDSummary
      summary: "dashboard counters: domains, ips, services, open_findings"
      x-required-role: viewer
      parameters:
        - {name: scopeID, in: path, required: true, schema: {type: string}}
      security:
        - session: []
        - token: []
        - proxy: []
      responses:
        "200": {description: ok}
        "400": {$ref: "#/components/responses/e400"}
        "401": {$ref: "#/components/responses/e401"}
        "403": {$ref: "#/components/responses/e403"}
        "404": {$ref: "#/components/responses/e404"}
        "409": {$ref: "#/components/responses/e409"}
  /api/v1/scopes/{scopeID}/targets:
    get:
      operationId: getScopesByScopeIDTargets
      summary: "?tag= filters; each entry carries its group_id"
      x-required-role: viewer
      parameters:
        - {name: scopeID, in: path, required: true, schema: {type: string}}
      security:
        - session: []
        - token: []
        - proxy: []
      responses:
        "200": {description: ok}
        "400": {$ref: "#/components/responses/e400"}
        "401": {$ref: "#/components/responses/e401"}
        "403": {$ref: "#/components/responses/e403"}
        "404": {$ref: "#/components/responses/e404"}
        "409": {$ref: "#/components/responses/e409"}
    post:
      operationId: postScopesByScopeIDTargets
      summary: "{value} or {values:[…]}, kind (domain, cidr, ip, asn — inferred if omitted), tags, mode (passive_only default, active, exclude), authorize: true to record active authorization, group to add them to a named group (created if needed; without it each value is a group of its own)"
      x-required-role: operator
      parameters:
        - {name: scopeID, in: path, required: true, schema: {type: string}}
      security:
        - session: []
        - token: []
        - proxy: []
      responses:
        "200": {description: ok}
        "201": {description: created}
        "400": {$ref: "#/components/responses/e400"}
        "401": {$ref: "#/components/responses/e401"}
        "403": {$ref: "#/components/responses/e403"}
        "404": {$ref: "#/components/responses/e404"}
        "409": {$ref: "#/components/responses/e409"}
  /api/v1/scopes/{scopeID}/target-groups:
    get:
      operationId: getScopesByScopeIDTargetGroups
      summary: "the company's target groups, each with its entries and authorized (every entry active with a recorded authorization)"
      x-required-role: viewer
      parameters:
        - {name: scopeID, in: path, required: true, schema: {type: string}}
      security:
        - session: []
        - token: []
        - proxy: []
      responses:
        "200": {description: ok}
        "400": {$ref: "#/components/responses/e400"}
        "401": {$ref: "#/components/responses/e401"}
        "403": {$ref: "#/components/responses/e403"}
        "404": {$ref: "#/components/responses/e404"}
        "409": {$ref: "#/components/responses/e409"}
    post:
      operationId: postScopesByScopeIDTargetGroups
      summary: "{name, values, tags, authorize} — one group from a list of domains, IPs and CIDRs; name defaults to the first value; 409 if the name is taken"
      x-required-role: operator
      parameters:
        - {name: scopeID, in: path, required: true, schema: {type: string}}
      security:
        - session: []
        - token: []
        - proxy: []
      responses:
        "200": {description: ok}
        "201": {description: created}
        "400": {$ref: "#/components/responses/e400"}
        "401": {$ref: "#/components/responses/e401"}
        "403": {$ref: "#/components/responses/e403"}
        "404": {$ref: "#/components/responses/e404"}
        "409": {$ref: "#/components/responses/e409"}
  /api/v1/scopes/{scopeID}/footprint:
    get:
      operationId: getScopesByScopeIDFootprint
      summary: "what the company owns, as counts: target groups, targets, names, hosts, services, runs, findings, screenshots, schedules, alert channels — and active_runs / live_fleets, which block a delete"
      x-required-role: viewer
      parameters:
        - {name: scopeID, in: path, required: true, schema: {type: string}}
      security:
        - session: []
        - token: []
        - proxy: []
      responses:
        "200": {description: ok}
        "400": {$ref: "#/components/responses/e400"}
        "401": {$ref: "#/components/responses/e401"}
        "403": {$ref: "#/components/responses/e403"}
        "404": {$ref: "#/components/responses/e404"}
        "409": {$ref: "#/components/responses/e409"}
  /api/v1/scan-params:
    get:
      operationId: getScanParams
      summary: "every tunable a run accepts, with type, range, default and help"
      x-required-role: viewer
      security:
        - session: []
        - token: []
        - proxy: []
      responses:
        "200": {description: ok}
        "400": {$ref: "#/components/responses/e400"}
        "401": {$ref: "#/components/responses/e401"}
        "403": {$ref: "#/components/responses/e403"}
        "404": {$ref: "#/components/responses/e404"}
        "409": {$ref: "#/components/responses/e409"}
  /api/v1/scopes/{scopeID}/scan-profiles:
    get:
      operationId: getScopesByScopeIDScanProfiles
      summary: "saved parameter presets"
      x-required-role: viewer
      parameters:
        - {name: scopeID, in: path, required: true, schema: {type: string}}
      security:
        - session: []
        - token: []
        - proxy: []
      responses:
        "200": {description: ok}
        "400": {$ref: "#/components/responses/e400"}
        "401": {$ref: "#/components/responses/e401"}
        "403": {$ref: "#/components/responses/e403"}
        "404": {$ref: "#/components/responses/e404"}
        "409": {$ref: "#/components/responses/e409"}
    post:
      operationId: postScopesByScopeIDScanProfiles
      summary: "{name, params, global, default}"
      x-required-role: operator
      parameters:
        - {name: scopeID, in: path, required: true, schema: {type: string}}
      security:
        - session: []
        - token: []
        - proxy: []
      responses:
        "200": {description: ok}
        "201": {description: created}
        "400": {$ref: "#/components/responses/e400"}
        "401": {$ref: "#/components/responses/e401"}
        "403": {$ref: "#/components/responses/e403"}
        "404": {$ref: "#/components/responses/e404"}
        "409": {$ref: "#/components/responses/e409"}
  /api/v1/scopes/{scopeID}/runs:
    get:
      operationId: getScopesByScopeIDRuns
      summary: "runs with progress counters; each run carries started_by (the account; a scheduled run's is the schedule's creator) and started_via (session, token, proxy or schedule)"
      x-required-role: viewer
      parameters:
        - {name: scopeID, in: path, required: true, schema: {type: string}}
      security:
        - session: []
        - token: []
        - proxy: []
      responses:
        "200": {description: ok}
        "400": {$ref: "#/components/responses/e400"}
        "401": {$ref: "#/components/responses/e401"}
        "403": {$ref: "#/components/responses/e403"}
        "404": {$ref: "#/components/responses/e404"}
        "409": {$ref: "#/components/responses/e409"}
    post:
      operationId: postScopesByScopeIDRuns
      summary: "start a run; the fields are under *Starting a run*"
      x-required-role: operator
      parameters:
        - {name: scopeID, in: path, required: true, schema: {type: string}}
      security:
        - session: []
        - token: []
        - proxy: []
      responses:
        "200": {description: ok}
        "201": {description: created}
        "400": {$ref: "#/components/responses/e400"}
        "401": {$ref: "#/components/responses/e401"}
        "403": {$ref: "#/components/responses/e403"}
        "404": {$ref: "#/components/responses/e404"}
        "409": {$ref: "#/components/responses/e409"}
  /api/v1/scopes/{scopeID}/schedules:
    get:
      operationId: getScopesByScopeIDSchedules
      summary: "[{profile, exit, vpn_config_id, pool_id, worker_count, every_hours, enabled, next_run_at, last_run_id, last_run_at, last_error, profile_id, params, wordlist_ids}]"
      x-required-role: viewer
      parameters:
        - {name: scopeID, in: path, required: true, schema: {type: string}}
      security:
        - session: []
        - token: []
        - proxy: []
      responses:
        "200": {description: ok}
        "400": {$ref: "#/components/responses/e400"}
        "401": {$ref: "#/components/responses/e401"}
        "403": {$ref: "#/components/responses/e403"}
        "404": {$ref: "#/components/responses/e404"}
        "409": {$ref: "#/components/responses/e409"}
    post:
      operationId: postScopesByScopeIDSchedules
      summary: "pool_id, worker_count, every_hours, start_at, profile_id, params, wordlist_ids, enabled} — targets narrows each run to those values, empty is every non-excluded target at the time — every_hours 1…8784 repeats from start_at (default now); 0 runs once at start_at`, then disables itself"
      x-required-role: operator
      parameters:
        - {name: scopeID, in: path, required: true, schema: {type: string}}
      security:
        - session: []
        - token: []
        - proxy: []
      responses:
        "200": {description: ok}
        "201": {description: created}
        "400": {$ref: "#/components/responses/e400"}
        "401": {$ref: "#/components/responses/e401"}
        "403": {$ref: "#/components/responses/e403"}
        "404": {$ref: "#/components/responses/e404"}
        "409": {$ref: "#/components/responses/e409"}
  /api/v1/runs/{runID}:
    delete:
      operationId: deleteRunsByRunID
      summary: "delete a finished run: tasks, targets, observations, fleet record and change events go with it, and its screenshots and raw output are removed from object storage — history that came from the run disappears. 409 while the run is going (stop it first) or its fleet is still being removed; answers…"
      x-required-role: operator
      parameters:
        - {name: runID, in: path, required: true, schema: {type: string}}
      security:
        - session: []
        - token: []
        - proxy: []
      responses:
        "200": {description: ok}
        "400": {$ref: "#/components/responses/e400"}
        "401": {$ref: "#/components/responses/e401"}
        "403": {$ref: "#/components/responses/e403"}
        "404": {$ref: "#/components/responses/e404"}
        "409": {$ref: "#/components/responses/e409"}
    get:
      operationId: getRunsByRunID
      summary: "{run, progress, fleet?} — fleet is present when the run has its own containers, and carries the reason if it is waiting or failed"
      x-required-role: viewer
      parameters:
        - {name: runID, in: path, required: true, schema: {type: string}}
      security:
        - session: []
        - token: []
        - proxy: []
      responses:
        "200": {description: ok}
        "400": {$ref: "#/components/responses/e400"}
        "401": {$ref: "#/components/responses/e401"}
        "403": {$ref: "#/components/responses/e403"}
        "404": {$ref: "#/components/responses/e404"}
        "409": {$ref: "#/components/responses/e409"}
  /api/v1/runs/{runID}/targets:
    get:
      operationId: getRunsByRunIDTargets
      summary: "per-target status, counters, skip reasons"
      x-required-role: viewer
      parameters:
        - {name: runID, in: path, required: true, schema: {type: string}}
      security:
        - session: []
        - token: []
        - proxy: []
      responses:
        "200": {description: ok}
        "400": {$ref: "#/components/responses/e400"}
        "401": {$ref: "#/components/responses/e401"}
        "403": {$ref: "#/components/responses/e403"}
        "404": {$ref: "#/components/responses/e404"}
        "409": {$ref: "#/components/responses/e409"}
  /api/v1/runs/{runID}/events:
    get:
      operationId: getRunsByRunIDEvents
      summary: "server-sent events stream; see the note under *Known gaps*"
      x-required-role: viewer
      parameters:
        - {name: runID, in: path, required: true, schema: {type: string}}
      security:
        - session: []
        - token: []
        - proxy: []
      responses:
        "200": {description: ok}
        "400": {$ref: "#/components/responses/e400"}
        "401": {$ref: "#/components/responses/e401"}
        "403": {$ref: "#/components/responses/e403"}
        "404": {$ref: "#/components/responses/e404"}
        "409": {$ref: "#/components/responses/e409"}
  /api/v1/runs/{runID}/activity:
    get:
      operationId: getRunsByRunIDActivity
      summary: "{tasks, stages, workers} — what is running where, right now. Each stage carries found and found_kind (names, addresses, open ports, web endpoints) and, for discovery stages, sources (names per source: subfinder:crtsh, shuffledns, seed); each task carries result with the same counts"
      x-required-role: viewer
      parameters:
        - {name: runID, in: path, required: true, schema: {type: string}}
      security:
        - session: []
        - token: []
        - proxy: []
      responses:
        "200": {description: ok}
        "400": {$ref: "#/components/responses/e400"}
        "401": {$ref: "#/components/responses/e401"}
        "403": {$ref: "#/components/responses/e403"}
        "404": {$ref: "#/components/responses/e404"}
        "409": {$ref: "#/components/responses/e409"}
  /api/v1/runs/{runID}/diff:
    get:
      operationId: getRunsByRunIDDiff
      summary: "change events this run produced: {kind, asset_kind, asset_id, before, after, created_at}"
      x-required-role: viewer
      parameters:
        - {name: runID, in: path, required: true, schema: {type: string}}
      security:
        - session: []
        - token: []
        - proxy: []
      responses:
        "200": {description: ok}
        "400": {$ref: "#/components/responses/e400"}
        "401": {$ref: "#/components/responses/e401"}
        "403": {$ref: "#/components/responses/e403"}
        "404": {$ref: "#/components/responses/e404"}
        "409": {$ref: "#/components/responses/e409"}
  /api/v1/runs/{runID}/footprint:
    get:
      operationId: getRunsByRunIDFootprint
      summary: "what the run owns, as counts: {tasks, targets, service_observations, screenshots, resolution_records, finding_observations, change_events, has_fleet} — what deleting it removes"
      x-required-role: viewer
      parameters:
        - {name: runID, in: path, required: true, schema: {type: string}}
      security:
        - session: []
        - token: []
        - proxy: []
      responses:
        "200": {description: ok}
        "400": {$ref: "#/components/responses/e400"}
        "401": {$ref: "#/components/responses/e401"}
        "403": {$ref: "#/components/responses/e403"}
        "404": {$ref: "#/components/responses/e404"}
        "409": {$ref: "#/components/responses/e409"}
  /api/v1/scopes/{scopeID}/domains:
    get:
      operationId: getScopesByScopeIDDomains
      summary: "?q= substring"
      x-required-role: viewer
      parameters:
        - {name: scopeID, in: path, required: true, schema: {type: string}}
      security:
        - session: []
        - token: []
        - proxy: []
      responses:
        "200": {description: ok}
        "400": {$ref: "#/components/responses/e400"}
        "401": {$ref: "#/components/responses/e401"}
        "403": {$ref: "#/components/responses/e403"}
        "404": {$ref: "#/components/responses/e404"}
        "409": {$ref: "#/components/responses/e409"}
  /api/v1/scopes/{scopeID}/graph:
    get:
      operationId: getScopesByScopeIDGraph
      summary: "{nodes, edges} for the name→address map"
      x-required-role: viewer
      parameters:
        - {name: scopeID, in: path, required: true, schema: {type: string}}
      security:
        - session: []
        - token: []
        - proxy: []
      responses:
        "200": {description: ok}
        "400": {$ref: "#/components/responses/e400"}
        "401": {$ref: "#/components/responses/e401"}
        "403": {$ref: "#/components/responses/e403"}
        "404": {$ref: "#/components/responses/e404"}
        "409": {$ref: "#/components/responses/e409"}
  /api/v1/scopes/{scopeID}/hosts:
    get:
      operationId: getScopesByScopeIDHosts
      summary: "addresses with ASN, PTR, country, cloud"
      x-required-role: viewer
      parameters:
        - {name: scopeID, in: path, required: true, schema: {type: string}}
      security:
        - session: []
        - token: []
        - proxy: []
      responses:
        "200": {description: ok}
        "400": {$ref: "#/components/responses/e400"}
        "401": {$ref: "#/components/responses/e401"}
        "403": {$ref: "#/components/responses/e403"}
        "404": {$ref: "#/components/responses/e404"}
        "409": {$ref: "#/components/responses/e409"}
  /api/v1/scopes/{scopeID}/hostrows:
    get:
      operationId: getScopesByScopeIDHostrows
      summary: "the Hosts table: one row per name→address pair, ?q= and ?unresolved=true to include names that resolve to nothing. Returns {rows, unresolved_hidden}; each row's sources says how the name was found (seed, subfinder:<provider>, shuffledns, dns)"
      x-required-role: viewer
      parameters:
        - {name: scopeID, in: path, required: true, schema: {type: string}}
      security:
        - session: []
        - token: []
        - proxy: []
      responses:
        "200": {description: ok}
        "400": {$ref: "#/components/responses/e400"}
        "401": {$ref: "#/components/responses/e401"}
        "403": {$ref: "#/components/responses/e403"}
        "404": {$ref: "#/components/responses/e404"}
        "409": {$ref: "#/components/responses/e409"}
...